// src/types/apiTypes.ts
/* eslint-disable camelcase */

export type Coordinates = {
    lon: number;
    lat: number;
}

export type Weather = {
    id: number;
    main: string;
    description: string;
    icon: string;
}

export type WeatherMain = {
    temp: number;
    feels_like: number;
    temp_min: number;
    temp_max: number;
    pressure: number;
    humidity: number;
}

export type Wind = {
    speed: number;
    deg: number;
}

export type Clouds = {
    all: number;
}

export type Sys = {
    type: number;
    id: number;
    country: string;
    sunrise: number;
    sunset: number;
}

export interface IWeatherData{
    coord: Coordinates;
    weather: Weather[];
    base: string;
    main: WeatherMain;
    visibility: number;
    wind: Wind;
    clouds: Clouds;
    dt: number;
    sys: Sys;
    timezone: number;
    id: number;
    name: string;
    cod: number;
}